+++ src/sectors/sectors.module.ts
@@ -2,11 +2,11 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { SectorsController } from './sectors.controller';
 import { SectorsService } from './sectors.service';
-import { Sector, Collaborator, Checklist, ServiceOrder } from '../entities';
+import { Sector, Collaborator, Checklist, ServiceOrder, Team } from '../entities';
 
 @Module({
   imports: [
-    TypeOrmModule.forFeature([Sector, Collaborator, Checklist, ServiceOrder]),
+    TypeOrmModule.forFeature([Sector, Collaborator, Checklist, ServiceOrder, Team]),
   ],
   controllers: [SectorsController],
   providers: [SectorsService],