+++ src/teams/dto/create-team.dto.ts
@@ -29,4 +29,10 @@ export class CreateTeamDto {
   @ArrayUnique()
   @IsUUID('4', { each: true })
   contractIds: string[];
+
+  @IsOptional()
+  @IsArray()
+  @ArrayUnique()
+  @IsUUID('4', { each: true })
+  sectorIds?: string[];
 }